requirements.txt
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
importlib-metadata==6.1.0
intersystems-iris @ https://raw.githubusercontent.com/intersystems-community/iris-driver-distribution/main/DB-API/intersystems_iris-3.3.1-py3-none-any.whl
intersystems-irispython @ https://raw.githubusercontent.com/intersystems-community/iris-driver-distribution/main/DB-API/intersystems_irispython-3.2.0-py3-none-any.whl
iris-dollar-list==0.9.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
Werkzeug==2.2.3
zipp==3.15.0
